Refactor response formatting for consistency across multiple commands
- Updated the Format methods in various response structs to use a consistent string formatting style, enhancing readability and maintainability.
- Removed unnecessary line breaks and improved the structure of output strings for better clarity.
- Adjusted helper functions to support new formatting requirements, ensuring uniformity in response outputs.

+++ types_command.go
@@ -43,7 +43,7 @@ type ResponseError struct {
 
 // Error implements the error interface
 func (e *ResponseError) Error() string {
-	return fmt.Sprintf("cc: %#02x, description: %s", e.completionCode, e.description)
+	return fmt.Sprintf("cc: %#02x, description: %s", uint8(e.completionCode), e.description)
 }
 
 func (e *ResponseError) CompletionCode() CompletionCode {
